refactor using of short model names
